+++ include/seqan3/search/fm_index/fm_index_cursor.hpp
@@ -27,12 +27,6 @@
 #include <seqan3/search/fm_index/fm_index.hpp>
 #include <seqan3/std/ranges>
 
-// namespace seqan3::detail
-// {
-//     // forward declaration
-//     auto get_suffix_array_range(fm_index_cursor<index_t> const & it);
-// } // namespace seqan3::detail
-
 namespace seqan3
 {
 
@@ -63,7 +57,6 @@ namespace seqan3
 template <typename index_t>
 class fm_index_cursor
 {
-
 public:
 
     /*!\name Member types
@@ -75,7 +68,7 @@ public:
     using size_type = typename index_type::size_type;
     //!\}
 
-protected:
+private:
     //!\privatesection
 
     /*!\name Member types
@@ -159,16 +152,16 @@ public:
     //!\brief Default constructor. Accessing member functions on a default constructed object is undefined behavior.
     //        Default construction is necessary to make this class semi-regular and e.g., to allow construction of
     //        std::array of iterators.
-    fm_index_cursor() noexcept = default;                                    //!< Default constructor.
-    fm_index_cursor(fm_index_cursor const &) noexcept = default;             //!< Copy constructor.
-    fm_index_cursor & operator=(fm_index_cursor const &) noexcept = default; //!< Copy assignment.
-    fm_index_cursor(fm_index_cursor &&) noexcept = default;                  //!< Move constructor.
-    fm_index_cursor & operator=(fm_index_cursor &&) noexcept = default;      //!< Move assignment.
-    ~fm_index_cursor() = default;                                            //!< Destructor.
+    fm_index_cursor() noexcept = default;                                    //!< Defaulted.
+    fm_index_cursor(fm_index_cursor const &) noexcept = default;             //!< Defaulted.
+    fm_index_cursor & operator=(fm_index_cursor const &) noexcept = default; //!< Defaulted.
+    fm_index_cursor(fm_index_cursor &&) noexcept = default;                  //!< Defaulted.
+    fm_index_cursor & operator=(fm_index_cursor &&) noexcept = default;      //!< Defaulted.
+    ~fm_index_cursor() = default;                                            //!< Defaulted.
 
     //! \brief Construct from given index.
     fm_index_cursor(index_t const & _index) noexcept : index(&_index), node({0, _index.index.size() - 1, 0, 0}),
-                                                       sigma(_index.index.sigma - index_t::is_collection_)
+                                                       sigma(_index.index.sigma - index_t::text_layout_mode)
     {}
     //\}
 
@@ -267,11 +260,13 @@ public:
      *
      * No-throw guarantee.
      */
-    template <Alphabet char_t>
+    template <typename char_t>
     bool extend_right(char_t const c) noexcept
     {
+        static_assert(std::ConvertibleTo<char_t, typename index_t::char_type>,
+                     "The character must be convertible to the alphabet of the index.");
+
         assert(index != nullptr);
-        assert(index->sigma == alphabet_size<char_t>);
 
         size_type _lb = node.lb, _rb = node.rb;
 
@@ -307,8 +302,10 @@ public:
     bool extend_right(seq_t && seq) noexcept
     {
         static_assert(std::ranges::ForwardRange<seq_t>, "The query must model ForwardRange.");
+        static_assert(std::ConvertibleTo<innermost_value_type_t<seq_t>, typename index_t::char_type>,
+                     "The alphabet of the sequence must be convertible to the alphabet of the index.");
+
         assert(index != nullptr); // range must not be empty!
-        assert(index->sigma == alphabet_size<innermost_value_type_t<seq_t>>);
 
         size_type _lb = node.lb, _rb = node.rb;
         size_type new_parent_lb = parent_lb, new_parent_rb = parent_rb;
@@ -445,13 +442,14 @@ public:
     template <std::ranges::Range text_t>
     auto path_label(text_t && text) const noexcept
     //!\cond
-        requires !index_t::is_collection_
+        requires index_t::text_layout_mode == text_layout::single
     //!\endcond
     {
         static_assert(std::ranges::InputRange<text_t>, "The text must model InputRange.");
-        static_assert(!(dimension_v<text_t> != 1), "The input cannot be a text collection.");
+        static_assert(dimension_v<text_t> == 1, "The input cannot be a text collection.");
+        static_assert(std::Same<innermost_value_type_t<text_t>, typename index_t::char_type>,
+                      "The alphabet types of the given text and index differ.");
         assert(index != nullptr);
-        assert(index->sigma == alphabet_size<value_type_t<text_t>>);
 
         size_type const query_begin = offset() - index->index[node.lb];
         return text | view::slice(query_begin, query_begin + query_length());
@@ -461,13 +459,14 @@ public:
     template <std::ranges::Range text_t>
     auto path_label(text_t && text) const noexcept
     //!\cond
-        requires index_t::is_collection_
+        requires index_t::text_layout_mode == text_layout::collection
     //!\endcond
     {
         static_assert(std::ranges::InputRange<text_t>, "The text collection must model InputRange.");
-        static_assert(!(dimension_v<text_t> != 2), "The input must be a text collection.");
+        static_assert(dimension_v<text_t> == 2, "The input must be a text collection.");
+        static_assert(std::Same<innermost_value_type_t<text_t>, typename index_t::char_type>,
+                      "The alphabet types of the given text and index differ.");
         assert(index != nullptr);
-        assert(index->sigma == alphabet_size<innermost_value_type_t<text_t>>);
 
         size_type const loc = offset() - index->index[node.lb];
         size_type const query_begin = loc - index->text_begin_rs.rank(loc + 1) + 1; // Substract delimiters
@@ -505,7 +504,7 @@ public:
      */
     std::vector<size_type> locate() const
     //!\cond
-        requires !index_t::is_collection_
+        requires index_t::text_layout_mode == text_layout::single
     //!\endcond
     {
         assert(index != nullptr);
@@ -521,7 +520,7 @@ public:
     //!\overload
     std::vector<std::pair<size_type, size_type>> locate() const
     //!\cond
-        requires index_t::is_collection_
+        requires index_t::text_layout_mode == text_layout::collection
     //!\endcond
     {
         assert(index != nullptr);
@@ -552,7 +551,7 @@ public:
      */
     auto lazy_locate() const
     //!\cond
-        requires !index_t::is_collection_
+        requires index_t::text_layout_mode == text_layout::single
     //!\endcond
     {
         assert(index != nullptr);
@@ -564,7 +563,7 @@ public:
     //!\overload
     auto lazy_locate() const
     //!\cond
-        requires index_t::is_collection_
+        requires index_t::text_layout_mode == text_layout::collection
     //!\endcond
     {
         assert(index != nullptr);