+++ test/unit/search/fm_index/bi_fm_index_aa27_test.cpp
@@ -5,10 +5,10 @@
 // shipped with this file and also available at: https://github.com/seqan/seqan3/blob/master/LICENSE.md
 // -----------------------------------------------------------------------------------------------------
 
-#include "fm_index_test_template.hpp"
 #include "fm_index_collection_test_template.hpp"
+#include "fm_index_test_template.hpp"
 
-using t1 = std::pair<bi_fm_index<text_layout::single>, std::vector<aa27>>;
+using t1 = std::pair<bi_fm_index<aa27, text_layout::single>, std::vector<aa27>>;
 INSTANTIATE_TYPED_TEST_CASE_P(aa27, fm_index_test, t1);
-using t2 = std::pair<bi_fm_index<text_layout::collection>, std::vector<std::vector<aa27>>>;
+using t2 = std::pair<bi_fm_index<aa27, text_layout::collection>, std::vector<std::vector<aa27>>>;
 INSTANTIATE_TYPED_TEST_CASE_P(aa27_collection, fm_index_collection_test, t2);